Many tests. A few known failures.
The failures are all because there should be 24 solved states, but the code
only accounts for 6. The 24 solved states are the 6 ways the white face can be
pointing, times 4 rotations (6*4=24).

Rather than add the rest of the 24 test cases, the plan is to rewrite the cube
representation such that one corner (probably DFR) is fixed in place and need
not be stored at all. This will make determining if a cube is solved much
faster. It will also make heuristisc work much better later on (I was actually
writing code for heuristics when I ran into this problem).