Rewrite cube representation so UFL is fixed
This fixes the problems detailed in the description for commit 83112ba7fb. More
tests written, and all now pass. One limitation, however, is that the UFL cubie
cannot be swapped in the viewer.

The C code is a little simpler, because it stores one less cubie. The Python
code, however, is much more complicated, as it must convert between human
readable input/output and the C code's turns, which always happen with the UFL
cubie in the UFL slot.

For example, if we make the turn "U", the C representation effectively performs
"U y'". If we want to do "UR", we would have to give the C code "UB".

One way the Python code could be cleaned up is to add a library for
manipulating turn strings. A lot of the added computation could be moved there.